+++ src/cloudformation_dataclasses/skeleton/templates/__init__.py
@@ -1,1 +1,1 @@
-"""Skeleton templates for cfn-init."""
+"""Skeleton templates for cfn-dataclasses-init."""